CUDA error code: 700 in absorption function
it happens when i=213427 sampling=5000 for 16010, rollback to check if this persist with the previous codes.
Or this maybe the result_size int type

+++ AnACor/double/ray_tracing.c
@@ -2157,7 +2157,7 @@ int ray_tracing_gpu_overall_kernel(int32_t low, int32_t up,
                                    const float *scattering_vector_list, const float *omega_list,
                                    const float *raw_xray,
                                    const float *omega_axis, const float *kp_rotation_matrix,
-                                   int64_t len_result,
+                                   int len_result,
                                    float *voxel_size, float *coefficients,
                                    int8_t *label_list_1d, int *shape, int64_t full_iteration,
                                    int32_t store_paths, float *h_result_list, int *h_face, float *h_angles, float *h_python_result_list);
@@ -2742,11 +2742,11 @@ extern "C"
 
     float *ray_tracing_gpu_overall(int32_t low, int32_t up,
                                     int *coord_list,
-                                    int64_t len_coord_list,
+                                    int32_t len_coord_list,
                                     const float *scattering_vector_list, const float *omega_list,
                                     const float *raw_xray,
                                     const float *omega_axis, const float *kp_rotation_matrix,
-                                    int64_t len_result,
+                                    int32_t len_result,
                                     float *voxel_size, float *coefficients, int8_t ***label_list,
                                     int8_t *label_list_1d, int *shape, int32_t full_iteration,
                                     int32_t store_paths)
